requirements.txt
allure-pytest==2.8.19
allure-python-commons==2.8.19
apipkg==1.5
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
execnet==1.7.1
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-html==3.0.0
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
toml==0.10.2
urllib3==1.25.10
